Room item layout (#835)
